requirements.txt
asgiref==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-crispy-forms==1.10.0
django-filter==2.4.0
django-import-export==2.5.0
django-pandas==0.6.2
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.6
pandas==1.2.1
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
